Make Resque.enqueue return something deterministic
Currently Resque.enqueue returns an empty [] or the results of the
after_enqueue_hooks that may have been run. If the before_enqueue_hook rejected
the operation we get back a nil.

This commit simply changes enqueue to return nil on rejection and true if the
job is actually queued.